Refactor cancellation ownership for size ratchet
