+++ supervisor/owner_stop.py
@@ -30,7 +30,7 @@ import logging
 import pathlib
 import time
 from datetime import datetime, timezone
-from typing import Any, Dict
+from typing import Any, Dict, Optional
 
 from ouroboros.utils import append_jsonl, utc_now_iso
 
@@ -401,6 +401,128 @@ def _settle_descendants_hard(
 def owner_stop_control_id(intent: Dict[str, Any]) -> str:
     """Deterministic episode/control identity from the durable stop request."""
     return f"ownerstop:{str(intent.get('request_id') or '')}"
+
+
+def _mark_owner_stop_control_drained(
+    owner_ctx: Any, drive_root: Optional[pathlib.Path], task_id: str,
+) -> bool:
+    """Stamp the owner-stop finalize control's DELIVERY on the durable intent.
+
+    The intent lives on the CANONICAL data root (``budget_drive_root`` first;
+    a forked task's mailbox drive differs). Idempotent (first drain wins). A
+    failed stamp is retried ONCE; still unconfirmed, a typed forensic event
+    is appended and no extended budget is assumed: the sweep keeps the
+    request+outer-cap deadline, and ``_owner_stop_window_elapsed`` reads the
+    same unstamped intent, bounding the worker by that anchor."""
+    try:
+        from ouroboros.cancel_intents import (
+            mark_finalize_control_drained,
+            resolve_owner_stop_intent,
+        )
+
+        root = (
+            str(getattr(owner_ctx, "budget_drive_root", "") or "")
+            or (str(drive_root) if drive_root is not None else "")
+        )
+        if not (root and task_id):
+            return False
+        root_path = pathlib.Path(root)
+        for _ in range(2):
+            if mark_finalize_control_drained(root_path, task_id):
+                return True
+            _intent_task_id, row = resolve_owner_stop_intent(root_path, task_id)
+            if isinstance(row, dict) and str(row.get("control_drained_at") or ""):
+                return True  # already stamped: the durable anchor is confirmed
+            if not isinstance(row, dict) or not row:
+                # The same stop request was hardened/settled while this
+                # mailbox row was being drained.  It is a stale control, not a
+                # persistence failure and must not buy a final model turn.
+                return False
+
+        append_jsonl(root_path / "logs" / "events.jsonl", {
+            "ts": utc_now_iso(), "type": "owner_stop_stamp_failed",
+            "task_id": task_id,
+        })
+        return False
+    except Exception:
+        log.debug("owner-stop drain stamp failed for %s", task_id, exc_info=True)
+        return False
+
+
+def _owner_stop_control_is_current(
+    owner_ctx: Any,
+    drive_root: Optional[pathlib.Path],
+    task_id: str,
+    control_msg_id: str,
+) -> bool:
+    """Whether this deterministic mailbox control still names the live policy."""
+    try:
+        from ouroboros.cancel_intents import resolve_owner_stop_intent
+
+        root = (
+            str(getattr(owner_ctx, "budget_drive_root", "") or "")
+            or (str(drive_root) if drive_root is not None else "")
+        )
+        if not (root and task_id and control_msg_id):
+            return False
+        _intent_task_id, intent = resolve_owner_stop_intent(
+            pathlib.Path(root), task_id,
+        )
+        return bool(intent) and owner_stop_control_id(intent) == str(control_msg_id)
+    except Exception:
+        log.debug(
+            "owner-stop control validation failed for %s",
+            task_id,
+            exc_info=True,
+        )
+        return False
+
+
+def _narrow_round_deadline(ctx: Any, candidate: Any) -> Optional[float]:
+    """Apply a later control deadline without ever widening task authority."""
+    try:
+        deadline = float(candidate)
+    except (TypeError, ValueError, OverflowError):
+        return ctx.deadline_ts
+    current = ctx.deadline_ts
+    if current is None:
+        ctx.deadline_ts = deadline
+    else:
+        try:
+            ctx.deadline_ts = min(float(current), deadline)
+        except (TypeError, ValueError, OverflowError):
+            ctx.deadline_ts = deadline
+    return ctx.deadline_ts
+
+
+def _owner_stop_window_elapsed(ctx: Any) -> bool:
+    """Bind and check the owner-stop deadline."""
+    try:
+        from ouroboros import task_pacing
+        from ouroboros.cancel_intents import (
+            STOP_POLICY_FINALIZE,
+            resolve_owner_stop_intent,
+            stop_policy,
+        )
+
+        root = getattr(ctx, "status_drive_root", None) or ctx.drive_root
+        if root is None or not ctx.task_id:
+            return False
+        _intent_task_id, intent = resolve_owner_stop_intent(
+            pathlib.Path(root), ctx.task_id,
+        )
+        if not isinstance(intent, dict) or stop_policy(intent) != STOP_POLICY_FINALIZE:
+            return False
+        deadline = owner_stop_deadline_ts(
+            intent, float(task_pacing.get_finalization_grace_sec()),
+        )
+        if deadline:
+            effective_deadline = _narrow_round_deadline(ctx, deadline)
+            return time.time() >= float(effective_deadline)
+        return True
+    except Exception:
+        log.debug("owner-stop window check failed for %s", ctx.task_id, exc_info=True)
+        return False
 
 
 def revoke_hardened_owner_stop_control(
